Improved child indexing with new method AnalysisResult::finishIndexChild
It is necessary to close the indexing of the child before continuing
with the parent analysis because addTriple does not have a AnalysisResult
parameter. That forces the client to maintain a stack in order to get
information about the depth of the added triple. This stack, however,
can only be correct if the child analysis is properly finished.

Since some plugins add properties to the child closing of the child
result cannot be done automatically but needs to be triggered by the
plugin.

In addition this commit fixes a few of the include paths used
throughout the lib. There are still many issues open where source
files contain global include statements like <strigi/analysisresult.h>
which means that rebuilding with changed public headers will always
fail. The compiler will find the (outdated) installed headers
instead of the local ones in the source tree.
A clean solution would be to change the source tree layout to not have
a dedicated "include" folder and always use local includes.