Fixed indexing of m3u files.
The analyzer now constructs absolute paths to the containing files and
creates nie:links relations to them. nie:hasLogicalPart does not make
much sense since the audio files are not actually contained in the
playlist, they are just referenced. NFO specifies the usage of rdf:list.
However, that is not supported in Nepomuk at the moment. Thus, we
cannot put any order on the linked audio files yet.

REVIEW: 103961