fix parsing of genre field in id3v2 tags and clean code up a bit
the genre field of a id3v2 tag might or might not be a number that matches
entries in a lookup table or alternatively a string. If it is a number,
then it might or might not be in parenthesis.
Handle all of the above and also handle the fact that some people might
enjoy adding numbers that are outside the range of the lookup table